pyproject.toml
[tool.poetry]
name = "hydra-optuna-example"
version = "0.1.0"
description = ""
authors = ["hppRC <hpp.ricecake@gmail.com>"]
readme = "readme"

[tool.poetry.dependencies]
python = "^3.9"
hydra-core = "^1.1.0"
omegaconf = "^2.1.0"
hydra-optuna-sweeper = "^1.1.0"
hydra-joblib-launcher = "^1.1.5"
hydra-colorlog = "^1.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
